[fix] Fixed test_organization_default_label breaking integration tests
The test "test_organization_default_label" depends on a test app
which is not part of openwisp-users, therefore this test should
be shipped in the test app rather than openwisp-users, because
the tests of openwisp-users are imported and re-run in other
modules as integration tests, which test_organization_default_label
was breaking.
